requirements.txt
accelerate==1.10.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
APScheduler==3.11.0
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.5
billiard==4.2.1
cachetools==6.2.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docling==2.50.0
docling-core==2.47.0
docling-ibm-models==3.9.1
docling-parse==4.2.3
easyocr==1.7.2
ecdsa==0.19.1
effdet==0.4.1
email-validator==2.3.0
emoji==2.15.0
et_xmlfile==2.0.0
faiss-cpu==1.13.2
Faker==37.6.0
fastapi==0.116.1
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-vision==3.11.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.2.4
groq==0.31.1
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
hiredis==3.2.1
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
kombu==5.5.4
langchain==0.3.27
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.5
langchain-deepseek==0.1.4
langchain-groq==0.3.8
langchain-ollama==1.0.1
langchain-openai==0.3.32
langchain-text-splitters==1.1.0
langdetect==1.0.9
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.24
latex2mathml==3.78.1
lazy_loader==0.4
limits==5.5.0
lxml==5.4.0
Mako==1.3.10
markdown-it-py==4.0.0
marko==2.2.0
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
networkx==3.5
ninja==1.13.0
nltk==3.9.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
olefile==0.47
ollama==0.6.1
omegaconf==2.3.0
onnx==1.20.0
onnxruntime==1.23.2
openai==1.106.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.2
pdf2image==1.17.0
pdfminer.six==20251228
pgvector==0.4.1
pi_heif==1.1.1
pikepdf==10.1.0
pillow==11.3.0
pluggy==1.6.0
polyfactory==2.22.2
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycocotools==2.0.11
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pylatexenc==2.10
pyngrok==7.4.0
pyparsing==3.2.5
pypdf==6.5.0
pypdfium2==4.30.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.11.16
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.14.3
redis==5.2.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
rtree==1.4.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.1
semchunk==2.2.2
sentence-transformers==5.1.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.47.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.8.28
tiktoken==0.11.0
timm==1.0.22
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.0
triton==3.4.0
typer==0.16.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
unstructured==0.18.21
unstructured-client==0.42.6
unstructured.pytesseract==0.3.15
unstructured_inference==1.1.2
uritemplate==4.2.0
urllib3==2.5.0
uuid_utils==0.12.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.17.3
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.22.0
zstandard==0.24.0
